+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/DidCommHttpServerError.scala
@@ -3,6 +3,7 @@ package io.iohk.atala.agent.server
 sealed trait DidCommHttpServerError
 
 object DidCommHttpServerError {
+  case class InvalidContentTypeError(error: String) extends DidCommHttpServerError
   case class RequestBodyParsingError(error: String) extends DidCommHttpServerError
   case class DIDCommMessageParsingError(error: String) extends DidCommHttpServerError
 }