+++ prism-agent/service/server/src/main/scala/io/iohk/atala/system/controller/SystemEndpoints.scala
@@ -3,11 +3,26 @@ package io.iohk.atala.system.controller
 import io.iohk.atala.api.http.EndpointOutputs.*
 import io.iohk.atala.api.http.{ErrorResponse, RequestContext}
 import io.iohk.atala.system.controller.http.HealthInfo
+import sttp.apispec.Tag
 import sttp.tapir.ztapir.stringBody
 import sttp.tapir.json.zio.jsonBody
 import sttp.tapir.*
 
 object SystemEndpoints {
+
+  private val tagName = "System"
+  private val tagDescription =
+    s"""
+       |The __${tagName}__ is a REST API that allows to check the system health and scrap the runtime metrics.
+       |
+       |The __health__ endpoint returns the current version of the running service.
+       |This information can be used to check the health status of the running service in the docker or kubernetes environment.
+       |
+       |The __metrics__ endpoint returns the runtime metrics of the running service scraped from the internal prometheus registry.
+       |This information is collected by the prometheus server and can be used to monitor the running service.
+       |""".stripMargin
+
+  val tag = Tag(tagName, Some(tagDescription))
 
   val health: PublicEndpoint[
     (RequestContext),
@@ -20,8 +35,8 @@ object SystemEndpoints {
       .in("_system" / "health")
       .out(jsonBody[HealthInfo].description("The health info object."))
       .errorOut(basicFailures)
-      .tag("System")
-      .summary("As a system user, check the health status of the running service")
+      .tag(tagName)
+      .summary("Check the health status of the running service")
       .description("Returns the health info object of the running service")
       .name("systemHealth")
 
@@ -34,11 +49,11 @@ object SystemEndpoints {
     endpoint.get
       .in(extractFromRequest[RequestContext](RequestContext.apply))
       .in("_system" / "metrics")
-      .out(stringBody.description("The metrics as pain strings."))
+      .out(stringBody.description("The metrics as plain strings."))
       .errorOut(basicFailures)
-      .tag("System")
-      .summary("As a system user, check the health status of the running service")
-      .description("Returns the health info object of the running service")
+      .tag(tagName)
+      .summary("Collect the runtime metrics of the running service")
+      .description("Returns the metrics of the running service from the internal Prometheus registry")
       .name("systemMetrics")
 
 }